requirements.txt
Flask
pycryptodome
requests
PyYAML
redis
ipaddress
